+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Creating initial version of CLI to monitor running scripts.
 
+- CLI works both for running and attaching to running processes via lldb.
+
+- Created shell plugin to monitor jobs w/ shell syntax.
+
 ## [0.0.7] - 2023-04-19
 
 - Update plugin version spec to use a tilde because caret isn't working for some reason